+++ backend/src/main/java/com/bicho/bet/bicho/bet/aposta/TipoAposta.java
@@ -1,0 +1,12 @@
+package com.bicho.bet.bicho.bet.aposta;
+
+public enum TipoAposta {
+    GRUPO, 
+    DUQUE,
+    TERNO,
+    QUADRA,
+    QUINA,
+    DEZENA,
+    CENTENA,
+    MILHAR,
+}
