+++ backend/src/main/java/com/bicho/bet/bicho/bet/core/CustomQuerydslPredicateExecutor.java
@@ -1,0 +1,11 @@
+package com.bicho.bet.bicho.bet.core;
+
+import com.querydsl.core.types.Predicate;
+import org.springframework.data.querydsl.QuerydslPredicateExecutor;
+
+import java.util.List;
+
+public interface CustomQuerydslPredicateExecutor<T> extends QuerydslPredicateExecutor<T> {
+    @Override
+    List<T> findAll(Predicate predicate);
+}
