--- backend/src/main/java/com/bicho/bet/bicho/bet/enums/StatusJogo.java
@@ -1,6 +1,0 @@
-package com.bicho.bet.bicho.bet.enums;
-
-public enum StatusJogo {
-    ABERTO,
-    FECHADO
-}
