--- backend/src/main/java/com/bicho/bet/bicho/bet/enums/TipoAposta.java
@@ -1,12 +1,0 @@
-package com.bicho.bet.bicho.bet.enums;
-
-public enum TipoAposta {
-    GRUPO, 
-    DUQUE,
-    TERNO,
-    QUADRA,
-    QUINA,
-    DEZENA,
-    CENTENA,
-    MILHAR,
-}
