--- backend/src/main/java/com/bicho/bet/bicho/bet/enums/TipoResultado.java
@@ -1,16 +1,0 @@
-package com.bicho.bet.bicho.bet.enums;
-
-public enum TipoResultado {
-    PRIMEIRA(1),
-    SEGUNDA(2),
-    TERCEIRA(3),
-    QUARTA(4),
-    QUINTA(5),
-    BICHO(0);
-
-    private final Integer value;
-
-    TipoResultado(Integer val) {
-        value = val;
-    }
-}
