+++ backend/src/main/java/com/bicho/bet/bicho/bet/jogo/JogoRepository.java
@@ -1,0 +1,10 @@
+package com.bicho.bet.bicho.bet.jogo;
+
+import com.bicho.bet.bicho.bet.core.CustomQuerydslPredicateExecutor;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface JogoRepository extends JpaRepository<Jogo, Long>, CustomQuerydslPredicateExecutor<Jogo> {
+    
+}
