+++ backend/src/main/java/com/bicho/bet/bicho/bet/jogo/StatusJogo.java
@@ -1,0 +1,6 @@
+package com.bicho.bet.bicho.bet.jogo;
+
+public enum StatusJogo {
+    ABERTO,
+    FECHADO
+}
