--- backend/src/main/java/com/bicho/bet/bicho/bet/models/conta/Conta.java
@@ -1,54 +1,0 @@
-package com.bicho.bet.bicho.bet.models.conta;
-
-import com.bicho.bet.bicho.bet.models.core.EntityId;
-
-import javax.persistence.Column;
-import javax.persistence.MappedSuperclass;
-
-@MappedSuperclass
-public abstract class Conta extends EntityId {
-    @Column(name = "nome")
-    private String nome;
-
-    @Column(name = "telefone")
-    private String telefone;
-
-    @Column(name = "saldo")
-    private double saldo;
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public String getTelefone() {
-        return telefone;
-    }
-
-    public void setTelefone(String telefone) {
-        this.telefone = telefone;
-    }
-
-    public double getSaldo() {
-        return saldo;
-    }
-
-    public void setSaldo(double saldo) {
-        this.saldo = saldo;
-    }
-
-    public void depositar(double valor) {
-        this.saldo += valor;
-    }
-
-    @Override
-    public String toString() {
-        return "Conta{" +
-                "nome='" + nome + '\'' +
-                ", saldo=" + saldo +
-                '}';
-    }
-}
