--- backend/src/main/java/com/bicho/bet/bicho/bet/models/core/EntityId.java
@@ -1,16 +1,0 @@
-package com.bicho.bet.bicho.bet.models.core;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
-
-@Getter
-@Setter
-@MappedSuperclass
-public class EntityId {
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    @Column(name = "id", nullable = false)
-    private Long id;
-}
