--- backend/src/main/java/com/bicho/bet/bicho/bet/repositories/ApostadorRepository.java
@@ -1,10 +1,0 @@
-package com.bicho.bet.bicho.bet.repositories;
-
-import com.bicho.bet.bicho.bet.models.conta.Apostador;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ApostadorRepository extends JpaRepository<Apostador, Long>, CustomQuerydslPredicateExecutor<Apostador> {
-    
-}
