+++ backend/src/main/java/com/bicho/bet/bicho/bet/resultado/ResultadoRepository.java
@@ -1,0 +1,11 @@
+package com.bicho.bet.bicho.bet.resultado;
+
+import com.bicho.bet.bicho.bet.core.CustomQuerydslPredicateExecutor;
+import com.bicho.bet.bicho.bet.resultado.Resultado;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ResultadoRepository extends JpaRepository<Resultado, Long>, CustomQuerydslPredicateExecutor<Resultado> {
+    
+}
